Move icon and label modules
