--- src/amulet_editor/models/widgets/__init__.py
@@ -1,6 +1,0 @@
-# from ._card import ALinkCard, QPixCard
-from ._icon import AStylableSvgWidget, AIconButton, ATooltipIconButton
-
-# from ._label import QElidedLabel
-# from ._menu import QMenuWidget
-# from ._line import QHLine
